// src/app/types/enums/ingredients.enum.ts
export enum Ingredient {
    TOMATO = 'TOMATO',
    MOZZARELLA = 'MOZZARELLA',
    HAM = 'HAM',
    OREGANO = 'OREGANO',
    GARLIC = 'GARLIC',
    ONION = 'ONION',
    MUSHROOMS = 'MUSHROOMS',
    PARMESAN = 'PARMESAN',
    BASIL = 'BASIL',
    PEPPERONI = 'PEPPERONI',
    EGG = 'EGG',
    CHILLI = 'CHILLI',
    SOUR_CREAM = 'SOUR_CREAM',
    BACON = 'BACON',
    BLUE_CHEESE = 'BLUE_CHEESE',
    GORGONZOLA = 'GORGONZOLA',
    TUNA = 'TUNA',
    OLIVES = 'OLIVES',
    CHEDDER = 'CHEDDER',
    CARAMELIZED_ONION = 'CARAMELIZED_ONION',
    BURGER_SAUCE = 'BURGER_SAUCE',
    ICEBERG = 'ICEBERG',
    CRUNCHY_BACON = 'CRUNCHY_BACON',
    GRILLED_ONION = 'GRILLED_ONION',
    PICKLES = 'PICKLES',
    GOUDA = 'GOUDA',
    SMOKED_BBQ = 'SMOKED_BBQ',
    DOUBLE_CHEESE = '2X_CHEESE',
    GRILLED_BACON = 'GRILLED_BACON',
    BBQ = 'BBQ',



}